Move code related to Junit report parsing into a dedicated package
