+++ pkg/junit/xml_test.go
@@ -1,10 +1,11 @@
-package report
+package junit_test
 
 import (
 	"encoding/xml"
 	"os"
 	"testing"
 
+	"github.com/radiofrance/dib/pkg/junit"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -14,12 +15,12 @@ func TestDIBReport_parseDgossLogs(t *testing.T) {
 	tests := []struct {
 		name     string
 		input    string
-		expected Testsuite
+		expected junit.Testsuite
 	}{
 		{
 			name:  "Goss tests succeed",
 			input: "../../test/fixtures/report/junit/junit-image-test.xml",
-			expected: Testsuite{
+			expected: junit.Testsuite{
 				XMLName:   xml.Name{Local: "testsuite"},
 				Name:      "goss",
 				Errors:    "0",
@@ -28,7 +29,7 @@ func TestDIBReport_parseDgossLogs(t *testing.T) {
 				Skipped:   "0",
 				Time:      "0.000",
 				Timestamp: "2022-10-20T18:29:26Z",
-				TestCases: []TestCase{
+				TestCases: []junit.TestCase{
 					{
 						XMLName:   xml.Name{Local: "testcase"},
 						ClassName: "goss-image-test",
@@ -51,7 +52,7 @@ func TestDIBReport_parseDgossLogs(t *testing.T) {
 		{
 			name:  "Some Goss tests failed",
 			input: "../../test/fixtures/report/junit/junit-image-test-fail.xml",
-			expected: Testsuite{
+			expected: junit.Testsuite{
 				XMLName:   xml.Name{Local: "testsuite"},
 				Name:      "goss",
 				Errors:    "0",
@@ -60,7 +61,7 @@ func TestDIBReport_parseDgossLogs(t *testing.T) {
 				Skipped:   "0",
 				Time:      "0.000",
 				Timestamp: "2022-10-20T18:29:26Z",
-				TestCases: []TestCase{
+				TestCases: []junit.TestCase{
 					{
 						XMLName:   xml.Name{Local: "testcase"},
 						ClassName: "goss-image-test",
@@ -90,7 +91,7 @@ func TestDIBReport_parseDgossLogs(t *testing.T) {
 
 			data, err := os.ReadFile(test.input)
 			assert.NoError(t, err)
-			actual, err := parseJunit(data)
+			actual, err := junit.ParseRawLogs(data)
 			assert.NoError(t, err)
 			assert.Equal(t, test.expected, actual)
 		})
